refactor: Namespace the Posts class within the Pinboard class This change also includes a drive-by fix to move the `Update` class into a separate file.
